Feat(test-common): Continue moving standalone test utilities to test-common
