+++ src/test/java/com/synopsys/integration/alert/provider/blackduck/collector/VulnerabilityMessageBuilderTest.java
@@ -40,7 +40,6 @@ import org.springframework.core.io.ClassPathResource;
 
 import com.google.common.io.Files;
 import com.google.gson.Gson;
-import com.synopsys.integration.alert.TestConstants;
 import com.synopsys.integration.alert.common.message.model.CommonMessageData;
 import com.synopsys.integration.alert.common.message.model.ComponentItem;
 import com.synopsys.integration.alert.common.message.model.LinkableItem;
@@ -49,6 +48,7 @@ import com.synopsys.integration.alert.common.persistence.accessor.FieldUtility;
 import com.synopsys.integration.alert.common.persistence.model.ConfigurationJobModel;
 import com.synopsys.integration.alert.common.util.DateUtils;
 import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
+import com.synopsys.integration.alert.provider.blackduck.BlackDuckMessageBuilderTestConstants;
 import com.synopsys.integration.alert.provider.blackduck.BlackDuckProperties;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.BlackDuckIssueTrackerCallbackUtility;
 import com.synopsys.integration.alert.provider.blackduck.collector.builder.MessageBuilderConstants;
@@ -94,7 +94,7 @@ public class VulnerabilityMessageBuilderTest {
         VulnerabilityNotificationMessageBuilder vulnMessageBuilder = new VulnerabilityNotificationMessageBuilder(blackDuckIssueTrackerCallbackUtility);
         ConfigurationJobModel job = Mockito.mock(ConfigurationJobModel.class);
         Mockito.when(job.getFieldUtility()).thenReturn(new FieldUtility(Map.of()));
-        VulnerabilityNotificationView notificationView = getNotificationView(TestConstants.VULNERABILITY_NOTIFICATION_JSON_PATH);
+        VulnerabilityNotificationView notificationView = getNotificationView(BlackDuckMessageBuilderTestConstants.VULNERABILITY_NOTIFICATION_JSON_PATH);
         CommonMessageData commonMessageData = new CommonMessageData(1L, 1L, "provider", "providerConfigName", "providerUrl", DateUtils.createCurrentDateTimestamp(), job);
         List<ProviderMessageContent> providerMessageContents = vulnMessageBuilder.buildMessageContents(commonMessageData, notificationView, blackDuckBucket, blackDuckServicesFactory);
 
@@ -125,7 +125,7 @@ public class VulnerabilityMessageBuilderTest {
         VulnerabilityNotificationMessageBuilder vulnMessageBuilder = new VulnerabilityNotificationMessageBuilder(blackDuckIssueTrackerCallbackUtility);
         ConfigurationJobModel job = Mockito.mock(ConfigurationJobModel.class);
         Mockito.when(job.getFieldUtility()).thenReturn(new FieldUtility(Map.of()));
-        VulnerabilityNotificationView notificationView = getNotificationView(TestConstants.VULNERABILITY_NOTIFICATION_JSON_PATH);
+        VulnerabilityNotificationView notificationView = getNotificationView(BlackDuckMessageBuilderTestConstants.VULNERABILITY_NOTIFICATION_JSON_PATH);
         CommonMessageData commonMessageData = new CommonMessageData(1L, 1L, "provider", "providerConfigName", "providerUrl", DateUtils.createCurrentDateTimestamp(), job);
         List<ProviderMessageContent> providerMessageContents = vulnMessageBuilder.buildMessageContents(commonMessageData, notificationView, blackDuckBucket, blackDuckServicesFactory);
 
@@ -159,7 +159,7 @@ public class VulnerabilityMessageBuilderTest {
         VulnerabilityNotificationMessageBuilder vulnMessageBuilder = new VulnerabilityNotificationMessageBuilder(blackDuckIssueTrackerCallbackUtility);
         ConfigurationJobModel job = Mockito.mock(ConfigurationJobModel.class);
         Mockito.when(job.getFieldUtility()).thenReturn(new FieldUtility(Map.of()));
-        VulnerabilityNotificationView notificationView = getNotificationView(TestConstants.VULNERABILITY_NOTIFICATION_JSON_PATH);
+        VulnerabilityNotificationView notificationView = getNotificationView(BlackDuckMessageBuilderTestConstants.VULNERABILITY_NOTIFICATION_JSON_PATH);
 
         CommonMessageData commonMessageData = new CommonMessageData(1L, 1L, "provider", "providerConfigName", "providerUrl", DateUtils.createCurrentDateTimestamp(), job);
         List<ProviderMessageContent> messageContents = vulnMessageBuilder.buildMessageContents(commonMessageData, notificationView, blackDuckBucket, blackDuckServicesFactory);
@@ -205,8 +205,8 @@ public class VulnerabilityMessageBuilderTest {
         Mockito.when(componentService.getRemediationInformation(Mockito.any(ComponentVersionView.class))).thenReturn(Optional.empty());
         Mockito.when(blackDuckServicesFactory.createComponentService()).thenReturn(componentService);
 
-        VulnerabilityNotificationView notification1 = getNotificationView(TestConstants.VULNERABILITY_NOTIFICATION_SIMPLE_01_JSON_PATH);
-        VulnerabilityNotificationView notification2 = getNotificationView(TestConstants.VULNERABILITY_NOTIFICATION_SIMPLE_02_JSON_PATH);
+        VulnerabilityNotificationView notification1 = getNotificationView(BlackDuckMessageBuilderTestConstants.VULNERABILITY_NOTIFICATION_SIMPLE_01_JSON_PATH);
+        VulnerabilityNotificationView notification2 = getNotificationView(BlackDuckMessageBuilderTestConstants.VULNERABILITY_NOTIFICATION_SIMPLE_02_JSON_PATH);
         List<VulnerabilityNotificationView> orderedNotifications = List.of(notification1, notification2);
 
         VulnerabilityNotificationMessageBuilder vulnMessageBuilder = new VulnerabilityNotificationMessageBuilder(blackDuckIssueTrackerCallbackUtility);
