+++ test-common/src/main/java/com/synopsys/integration/alert/util/TestAlertProperties.java
@@ -1,9 +1,32 @@
+/**
+ * test-common
+ *
+ * Copyright (c) 2020 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 package com.synopsys.integration.alert.util;
 
 import java.util.Optional;
 
 import com.synopsys.integration.alert.common.AlertProperties;
 
+// TODO rename class to include "Mock"
 public class TestAlertProperties extends AlertProperties {
     public static final String PROPERTY_USER_DIR = "user.dir";
     public static final String RESOURCES_PATH = "/src/main/resources";
@@ -92,4 +115,5 @@ public class TestAlertProperties extends AlertProperties {
     public void setAlertSecretsDir(String alertSecretsDir) {
         this.alertSecretsDir = alertSecretsDir;
     }
+
 }
